Change fully written error ZObject to a constructed one
In order to improve readability, using the error function
to create the error messages, and using the constants for
the ZIDs for the error messages.
I am not sure if this is better. It kinda makes it more
readable but one could argue that it requires the error
module to be correct, and thus takes away from the testability.

Bug: T271761
Change-Id: I8d225928692905e05d0bda44db35554fcd40462b