Add Tests and Functionality to Orchestrator for StrToChars()
Ensure that the StrToChars function (Z886) supports generic Typed Lists
as output.

With this patch, both the old test cases
(with Z10 lists used to specify the output)
and the new test cases
(with Z881 lists used to specify the output)
are passing.
All of these test cases are now returning Z881 lists.

Bug: T294482
Change-Id: I809154161ad7513aecdc1b1e1ddc24d8347a4813