Rename one of Z10006s to Z10005 to get rid of a conflict
Z10006 was used in both:
* test_data/Z10006.json (introduced in Ibaf1877) and
* wikilambda_fetch.json (introduced in I8c1c08e)

wikilambda_fetch.json should only contain bultins, so I want to move the
extra 3 ZObjects from it to test_data, thus the need for deduplication.

In a followup wikilambda_fetch.json will be replaced in favour of
loading builtins from `function-schemata`

And in general it's probably a good idea not to duplicate ZIDs to avoid
confusion and future bugs.

Bug: T300651
Change-Id: Ic3bfaef7f26a19c5883423fa601b99f83ac80989