Define a basic `.pipeline/config.yaml`
The previous `.pipeline/config.yaml` was ignored as Jenkins and Zuul had
not been configured to invoke a self-service pipeline config for this
project. Now that they do, provide a basic `.pipeline/config.yaml` that
constitutes the same operations as the previously used but now
deprecated service-pipeline-test and service-pipeline-test-and-publish
Jenkins jobs.

Bug: T307507
Change-Id: I8715c74844b293822b6d021800cad129a2f07770