Do not resolve implementations before selecting them
This bug was causing composition implementations to be evaluated every time they were retieved before they were selected. That was incorrect because the scope should not be populated with
arguments at that point (although it currently is because of TT309195). It also means that implementations were evaluated mutliple times, possibly leading to exponential explosion in the case of recursive functions (e.g. factorial).

I had to temporarily disable the curry test. It had been working until now because of a subtle interaction of multiple bugs (including T309195 and this one), and making it pass again requires fixing all of them at once.

Tested it also on the Sandbox API examples, they seem to be working.

Bug: T310995
Change-Id: I3407bf918f8e682ce95710adb4effcfc77c2ca4c