Fix ZWrapper caching by storing the result of argument reference resolution.
Not sure why it was changed this way in https://gerrit.wikimedia.org/r/c/mediawiki/services/function-orchestrator/+/805261/ but it broke the resolveKey functionality. The invariant around resolveKey is that the key should be resolved after it is called, but this was not the case for argument references.

I am rolling back some of the changes that were made there. I am also fixing the map test in another change to make it not depend on dynamic scoping.

Bug: T312227
Change-Id: I73fe263d92c3d31ac2f424d399b69ced90870de0