Add Scott numerals unit tests in the orchestrator.
These tests represent numbers using [Scott encoding](https://en.wikipedia.org/wiki/Mogensen%E2%80%93Scott_encoding) and define functions such as the Ackermann function using them.

Why represent numbers like that? Scott (and Church) encoding are fundamental to the lambda calculus and these examples exercise deep recursion and currying, so having them greatly increases our confidence that scoping is done correctly, at least at the computational level of the orchestrator (that is, not the code that does verification). Moreover, the examples demonstrate that the orchestrator is *really* slow at the moment; just computing Ackermann(2, 2) == 7 takes 15 seconds, while it should be nearly instantaneous for such low numbers. Given that this test is all in-process, this gives us a good benchmark to optimize the orchestrator.

To properly test that the result is correct, I define a conversion function using a "native" increment function (i.e. maps "2" to "3", "3" to "4", etc.) because that's the easiest way. For this I needed to modify the canned evaluator response to not return a constant value but use a callback instead.

Aside: Why Scott encoding instead of Church encoding? Scott encodings represent datatypes as pattern matching while Church encodings represent datatypes as folding. Both are fine, but Scott encoding is more natural in the sense that it is closer to the way we program in functional programming languages: using pattern matching and recursion.

Bug: T310093
Bug: T309195
Change-Id: I7461833a726dacc1e4753753f3c741f56ed4a4d2