Convert caught JS errors to ZErrors and return them
This handles errors caught in orchestrate().
This should be merged after function-schemata patch
861923 is available.

This patch set exposes a bug that
causes the failure of the orchestrator unit test
'Test that reify avoids infinite expansions'.
That bug needs to be fixed separately, so
this patch set is setting that test to be skipped.
The bug has been reported in T326518.

Bug: T321389
Change-Id: I3242e5e561c445e5c24d3c29df090d736ae7b87c