Add tests for multilingual text validator.
As discussed, it is not currently possible to call the orchestrator with
Z112 (if you provide a Z99 as input it complains that the input is of
wrong type; if you provide Z12 the built-in implementation fails). So
instead I am implicitly testing the validator by sending calls to other
functions that contain multilingual text.

Bug: T324845
Bug: T322956
Change-Id: I3fab2e1535119e25e093efbe485c10fdf2c4c356